+++ packages/spot-consumer/src/composers/index.ts
@@ -1,2 +1,1 @@
 export * from './SpotMarketComposer'
-export * from './SubaccountComposer'