+++ packages/spot-consumer/src/consumers/index.ts
@@ -1,3 +1,2 @@
-export * from './SubaccountConsumer'
 export * from './SpotMarketChronosConsumer'
 export * from './SpotMarketConsumer'