+++ packages/spot-consumer/src/streams/SpotMarket/SpotMarketOrderStream.ts
@@ -3,8 +3,9 @@ import {
   StreamMarketOrdersResponse,
 } from '@injectivelabs/exchange-api/injective_spot_markets_rpc_pb'
 import { InjectiveSpotMarketsRPCClient } from '@injectivelabs/exchange-api/injective_spot_markets_rpc_pb_service'
+import { StreamOperation } from '@injectivelabs/ts-types'
 import { SpotMarketTransformer } from '../../transformers/SpotMarketTransformer'
-import { StreamOperation, SpotMarketOrder } from '../../types'
+import { SpotMarketOrder } from '../../types'
 
 export type SpotMarketOrderStreamCallback = ({
   order,