+++ packages/spot-consumer/src/streams/index.ts
@@ -1,2 +1,1 @@
 export * from './SpotMarketStream'
-export * from './SubaccountStream'