+++ packages/spot-consumer/src/transformers/SpotMarketTransformer.ts
@@ -1,11 +1,10 @@
+import { TradeDirection, TradeExecutionType } from '@injectivelabs/ts-types'
 import {
   GrpcPriceLevel,
   GrpcSpotMarketInfo,
   GrpcSpotMarketOrder,
   GrpcSpotMarketTrade,
   SpotOrderType,
-  TradeDirection,
-  TradeExecutionType,
   Orderbook,
   PriceLevel,
   SpotMarket,