+++ packages/spot-consumer/src/transformers/index.ts
@@ -1,2 +1,1 @@
 export * from './SpotMarketTransformer'
-export * from './SubaccountTransformer'