+++ packages/spot-consumer/src/types/index.ts
@@ -1,16 +1,6 @@
-import { CosmosCoin as GrpcCosmosCoin } from '@injectivelabs/exchange-api/injective_accounts_rpc_pb'
 import { PriceLevel as GrpcPriceLevel } from '@injectivelabs/exchange-api/injective_spot_markets_rpc_pb'
 
 export * from './spot-market'
 export * from './spot-market-chronos'
-export * from './subaccount'
 
-export enum StreamOperation {
-  Insert = 'insert',
-  Delete = 'delete',
-  Replace = 'replace',
-  Update = 'update',
-  Invalidate = 'invalidate',
-}
-
-export { GrpcPriceLevel, GrpcCosmosCoin }
+export { GrpcPriceLevel }