+++ packages/subaccount-consumer/src/types/index.ts
@@ -1,0 +1,5 @@
+import { CosmosCoin as GrpcCosmosCoin } from '@injectivelabs/exchange-api/injective_accounts_rpc_pb'
+
+export * from './subaccount'
+
+export { GrpcCosmosCoin }