+++ packages/ts-types/src/cosmos.ts
@@ -1,0 +1,12 @@
+export interface Coin {
+  denom: string
+  amount: string
+}
+
+export enum StreamOperation {
+  Insert = 'insert',
+  Delete = 'delete',
+  Replace = 'replace',
+  Update = 'update',
+  Invalidate = 'invalidate',
+}