+++ packages/ts-types/src/index.ts
@@ -1,6 +1,8 @@
 export * from './enums'
 export * from './aliases'
 export * from './transactions'
+export * from './cosmos'
+export * from './trade'
 
 export interface Constructable<T> {
   new (...args: never): T