+++ packages/ts-types/src/trade.ts
@@ -1,0 +1,13 @@
+export enum TradeExecutionType {
+  Market = 'market',
+  LimitFill = 'limitFill',
+  LimitMatchRestingOrder = 'limitMatchRestingOrder',
+  LimitMatchNewOrder = 'limitMatchNewOrder',
+}
+
+export enum TradeDirection {
+  Buy = 'buy',
+  Sell = 'sell',
+  Long = 'long',
+  Short = 'short',
+}