+++ packages/exceptions/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.19](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/exceptions@0.0.18...@injectivelabs/exceptions@0.0.19) (2021-04-27)
+
+**Note:** Version bump only for package @injectivelabs/exceptions
+
 ## [0.0.18](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/exceptions@0.0.17...@injectivelabs/exceptions@0.0.18) (2021-03-14)
 
 **Note:** Version bump only for package @injectivelabs/exceptions