+++ packages/spot-consumer/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.23](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/spot-consumer@0.0.22...@injectivelabs/spot-consumer@0.0.23) (2021-04-27)
+
+### Features
+
+* subaccount decoupled from spot to own package ([d6a0875](https://github.com/InjectiveLabs/injective-ts/commit/d6a0875fe4a4c1dd6501dbea8dcf05d6bdd7d585))
+
 ## [0.0.22](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/spot-consumer@0.0.21...@injectivelabs/spot-consumer@0.0.22) (2021-04-26)
 
 ### Bug Fixes