+++ packages/ts-types/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.16](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/ts-types@0.0.15...@injectivelabs/ts-types@0.0.16) (2021-04-27)
+
+### Features
+
+* subaccount decoupled from spot to own package ([d6a0875](https://github.com/InjectiveLabs/injective-ts/commit/d6a0875fe4a4c1dd6501dbea8dcf05d6bdd7d585))
+
 ## [0.0.15](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/ts-types@0.0.14...@injectivelabs/ts-types@0.0.15) (2021-03-14)
 
 **Note:** Version bump only for package @injectivelabs/ts-types